autoload result helper file correctly
copy prewk/result's helpers.php into own Helpers directory, so we can autoload it properly without relying on vendor/ being there. This is not the case when you're in the project which depends on this library, composer will then look for another vendor/ directory inside the directory.

This debacle is avoided if we just copy the helpers file (as intended probably)